requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pytest
nltk==3.2.5
tensorboardX
requests_oauthlib
mecab-python3
neologdn
scikit-learn==0.19.1
